Move bodhi tests to their own module.
